+++ opt/scopeserver/dataserver/utils/data_file_handler.py
@@ -60,9 +60,7 @@ class DataFileHandler:
 
     def __init__(self):
         self.current_UUIDs = {}
-        self.hsap_to_dmel_mappings = None
-        self.mmus_to_dmel_mappings = None
-        self.dmel_mappings = None
+        self.dmel_mappings = {}
         self.permanent_UUIDs = set()
         self.orcid_ids = {}
         self.active_sessions = {}
@@ -74,6 +72,7 @@ class DataFileHandler:
         self.logs_dir = DataFileHandler.get_data_dir_path_by_file_type(file_type="Logs")
         self.create_global_dirs()
         self.create_uuid_log()
+        self.load_gene_mappings()
 
     def get_gene_sets_dir(self) -> str:
         return self.gene_sets_dir
@@ -230,12 +229,5 @@ class DataFileHandler:
 
     def load_gene_mappings(self) -> None:
         gene_mappings_dir_path = os.path.join(scopeserver.__path__[0], "dataserver", "data", "gene_mappings")  # type: ignore
-        DataFileHandler.dmel_mappings = pickle.load(
-            open(os.path.join(gene_mappings_dir_path, "terminal_mappings.pickle"), "rb")
-        )
-        DataFileHandler.hsap_to_dmel_mappings = pickle.load(
-            open(os.path.join(gene_mappings_dir_path, "hsap_to_dmel_mappings.pickle"), "rb")
-        )
-        DataFileHandler.mmus_to_dmel_mappings = pickle.load(
-            open(os.path.join(gene_mappings_dir_path, "mmus_to_dmel_mappings.pickle"), "rb")
-        )
+        with open(os.path.join(gene_mappings_dir_path, "terminal_mappings.pickle"), "rb") as tm_fh:
+            self.dmel_mappings = pickle.load(tm_fh)